# src/client/generated/w3s-openapi.yaml
openapi: 3.0.3
info:
  title: W3S
  description: All W3S APIs
  version: '1.0'
servers:
  - url: https://api.circle.com/v1/w3s
security:
  - BearerAuth: []
...
paths:
...
  /developer/walletSets:
    post:
      tags:
        - Wallet Sets
      summary: Create a new wallet set
      description: Generate a developer controlled wallet set.
      operationId: createWalletSet
      requestBody:
        required: true
        description: WalletSet creation request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeveloperWalletSetRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CreateDeveloperWalletSetResponse'
          description: WalletSet already exists
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDeveloperWalletSetResponse'
          description: WalletSet created
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /developer/walletSets/{id}:
    put:
      tags:
        - Wallet Sets
      summary: Update a wallet set
      description: Update the wallet set associated with the given wallet set ID.
      operationId: updateWalletSet
      parameters:
        - $ref: '#/components/parameters/WalletSetIdPath'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWalletSetMetadataRequest'
        description: Update walletSet metadata request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UpdateWalletSetMetadataResponse'
          description: WalletSet metadata updated
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /walletSets:
    get:
      tags:
        - Wallet Sets
      summary: List wallet sets
      description: Retreive a list of all wallet sets.
      operationId: listWalletSets
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/PageBefore'
        - $ref: '#/components/parameters/PageAfter'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GetWalletSetsResponse'
          description: WalletSets found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Link:
              description: 'Pagination cursor information, it can have four links depending on the page: self, first, next and prev. format: <https://api.circle.com/w3s/walletSets?to=2020-06-02T12:01:00Z&pageSize=50>; rel=\"next\'
              schema:
                type: string
        default:
          $ref: '#/components/responses/Default'
  /walletSets/{id}:
    get:
      tags:
        - Wallet Sets
      summary: Get a wallet set
      description: Retreive info for a single wallet set.
      operationId: getWalletSet
      parameters:
        - $ref: '#/components/parameters/WalletSetIdPath'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GetWalletSetByIDResponse'
          description: WalletSet found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /developer/wallets:
    post:
      tags:
        - Wallets
      summary: Create wallets
      description: Generates a new developer-controlled wallet or batch of wallets within a wallet set, specifying blockchain and wallet name.
      operationId: createDeveloperWallet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeveloperWalletRequest'
        description: Wallets creation request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CreateDeveloperWalletResponse'
          description: Wallets created
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /wallets:
    get:
      tags:
        - Wallets
      summary: List wallets
      description: Retrieves a list of all developer-controlled wallets that fit the specified parameters.
      operationId: listWallets
      parameters:
        - $ref: '#/components/parameters/XUserToken'
        - $ref: '#/components/parameters/Address'
        - $ref: '#/components/parameters/Blockchain'
        - $ref: '#/components/parameters/WalletSetIdQuery'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/PageBefore'
        - $ref: '#/components/parameters/PageAfter'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ExtGetAllWalletsResponse'
          description: Wallets found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Link:
              description: 'Pagination cursor information, it can have four links depending on the page: self, first, next and prev. format: <https://api.circle.com/w3s/wallets?to=2020-06-02T12:01:00Z&pageSize=50>; rel=\"next\'
              schema:
                type: string
        default:
          $ref: '#/components/responses/Default'
  /wallets/{id}:
    get:
      tags:
        - Wallets
      summary: Get a wallet
      description: Retrieves info for a single developer-controlled wallet using it's unique identifier.
      operationId: getWallet
      parameters:
        - name: id
          description: Wallet ID
          in: path
          required: true
          schema:
            format: uuid
            type: string
        - $ref: '#/components/parameters/XUserToken'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ExtGetWalletByIDResponse'
          description: Wallet found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
    put:
      tags:
        - Wallets
      summary: Update a wallet
      description: Updates info for a single developer-controlled wallet using it's unique identifier.
      operationId: updateWallet
      parameters:
        - name: id
          description: Wallet ID
          in: path
          required: true
          schema:
            format: uuid
            type: string
        - $ref: '#/components/parameters/XUserToken'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtUpdateWalletMetadataRequest'
        description: Wallet metadata update request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ExtWalletResponse'
          description: Wallet metadata updated
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /wallets/{id}/balances:
    get:
      tags:
        - Wallets
      summary: Get token balance for a wallet
      description: Fetches the digital asset balance for a single developer-controlled wallet using its unique identifier.
      operationId: listWalletBallance
      parameters:
        - description: Wallet ID
          in: path
          name: id
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XUserToken'
        - $ref: '#/components/parameters/IncludeAll'
        - $ref: '#/components/parameters/Name'
        - $ref: '#/components/parameters/TokenAddress'
        - $ref: '#/components/parameters/Standard'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/PageBefore'
        - $ref: '#/components/parameters/PageAfter'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GetBalancesResponse'
          description: Success
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /wallets/{id}/nfts:
    get:
      tags:
        - Wallets
      summary: Get NFTs for a wallet
      description: Fetches the info for all NFTs stored in a single developer-controlled wallet, using the wallets unique identifier.
      operationId: listWalletNfts
      parameters:
        - description: Wallet ID
          in: path
          name: id
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XUserToken'
        - $ref: '#/components/parameters/IncludeAll'
        - $ref: '#/components/parameters/Name'
        - $ref: '#/components/parameters/TokenAddress'
        - $ref: '#/components/parameters/Standard'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/PageBefore'
        - $ref: '#/components/parameters/PageAfter'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GetNftsResponse'
          description: Success
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
...
  /user/wallets:
    post:
      tags:
        - Wallets
      summary: Create wallets
      description: Generates a challenge for creating a new user-controlled wallet or batch of wallets, specifying blockchain and wallet name.
      operationId: createUserWallet
      parameters:
        - $ref: '#/components/parameters/XUserTokenRequired'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEndUserWalletRequest'
        description: walletResponse creation request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CreateEndUserWalletResponse'
          description: walletResponse already exists
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: integer
          description: Code is the error code
        message:
          type: string
          description: Message is the error message
...
    Blockchain:
      type: string
      enum:
        - ETH-GOERLI
        - ETH
        - AVAX-FUJI
        - AVAX
        - MATIC-MUMBAI
        - MATIC
    UpdateDate:
      type: string
      format: date-time
      description: Last update date of the resource. ISO-8601 UTC date/time.
      example: '2023-01-01T12:04:05Z'
...
    schemas_Id:
      type: string
      format: uuid
      description: Unique system generated identifier of the resource.
    TokenStandard:
      type: string
      enum:
        - ERC20
        - ERC721
        - ERC1155
    schemas_CreateDate:
      type: string
      format: date-time
      description: Date and time the resource was created. ISO-8601 UTC date/time.
      example: '2023-01-01T12:04:05Z'
    TokenResponse:
      type: object
      required:
        - id
        - blockchain
        - isNative
        - createDate
        - updateDate
      properties:
        id:
          $ref: '#/components/schemas/schemas_Id'
        name:
          type: string
          description: Blockchain name of the specified token.
        standard:
          $ref: '#/components/schemas/TokenStandard'
        blockchain:
          $ref: '#/components/schemas/Blockchain'
        decimals:
          type: integer
          description: Number of decimal places shown in the token amount.
        isNative:
          type: boolean
          description: Defines if the token is a native token of the specified blockchain. If TRUE, the token is a native token.
        symbol:
          type: string
          description: Blockchain symbol of the specified token.
        tokenAddress:
          type: string
          description: Blockchain address of the specified token.
        updateDate:
          $ref: '#/components/schemas/UpdateDate'
        createDate:
          $ref: '#/components/schemas/schemas_CreateDate'
...
    schemas_IdempotencyKey:
      type: string
      description: Universally unique identifier (UUID v4) idempotency key. This key is utilized to ensure exactly-once execution of mutating requests. To create a UUIDv4 go to [uuidgenerator.net](https://www.uuidgenerator.net).
      example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
      format: uuid
    EntitySecretCiphertext:
      type: string
      description: |
        A base64 string expression of the entity secret ciphertext. The entity secret should be encrypted by the entity public key. Circle mandates that the entity secret ciphertext is unique for each API request.
      format: byte
      example: 0NtD3d3+nmgb4GqYQXzAjKF8h5Zq6sHM2k/7UQz5OIXWnU7H8O9XvVYXqDnw1wEmY...
    CreateDeveloperWalletSetRequest:
      type: object
      required:
        - entitySecretCiphertext
        - idempotencyKey
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
        entitySecretCiphertext:
          $ref: '#/components/schemas/EntitySecretCiphertext'
        name:
          type: string
          example: treasury wallet set 1
    CustodyType:
      type: string
      enum:
        - DEVELOPER
        - ENDUSER
    ExternalUserId:
      type: string
      description: Unique system generated identifier for the user.
      example: ext_user_id_1
      maxLength: 50
      minLength: 5
    WalletSetResponse:
      type: object
      required:
        - id
        - custodyType
        - updateDate
        - createDate
      properties:
        id:
          $ref: '#/components/schemas/schemas_Id'
        createDate:
          $ref: '#/components/schemas/schemas_CreateDate'
        custodyType:
          $ref: '#/components/schemas/CustodyType'
        name:
          type: string
          description: Name or description of the wallet set.
          example: treasury wallet set 1
        updateDate:
          type: string
          format: date-time
          description: Last update date of the wallet set. ISO-8601 UTC date/time.
          example: '2023-01-01T12:04:05Z'
        userId:
          $ref: '#/components/schemas/ExternalUserId'
    CreateDeveloperWalletSetResponse:
      type: object
      properties:
        walletSet:
          $ref: '#/components/schemas/WalletSetResponse'
    UpdateWalletSetMetadataRequest:
      type: object
      properties:
        name:
          type: string
          example: My updated wallet set name
          minLength: 1
    UpdateWalletSetMetadataResponse:
      type: object
      properties:
        walletSet:
          $ref: '#/components/schemas/WalletSetResponse'
    GetWalletSetsResponse:
      type: object
      properties:
        walletSets:
          type: array
          items:
            $ref: '#/components/schemas/WalletSetResponse'
    GetWalletSetByIDResponse:
      type: object
      properties:
        walletSet:
          $ref: '#/components/schemas/WalletSetResponse'
    TransactionReferenceId:
      type: string
      example: grouptransaction123
      description: Optional reference or description used to identify the transaction.
    WalletMetadata:
      type: object
      properties:
        name:
          type: string
          description: Name or description associated with the wallet.
          example: My Wallet 1
        refId:
          $ref: '#/components/schemas/TransactionReferenceId'
    WalletSetId:
      type: string
      description: Unique system generated identifier of the wallet's parent wallet set. under.
      example: 71f2a6b4-ffa7-417a-ad5b-fb928753edc8
    CreateDeveloperWalletRequest:
      type: object
      required:
        - blockchains
        - entitySecretCiphertext
        - walletSetId
        - idempotencyKey
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
        blockchains:
          type: array
          items:
            $ref: '#/components/schemas/Blockchain'
          description: Blockchain(s) the requested wallets will be created on.
        count:
          type: integer
          description: Number of wallets that will be created per specified blockchain.
          example: 2
          maximum: 20
          minimum: 1
        entitySecretCiphertext:
          $ref: '#/components/schemas/EntitySecretCiphertext'
        metadata:
          type: array
          items:
            $ref: '#/components/schemas/WalletMetadata'
          description: List of metadata fields to associate with the corresponding wallet. If count is specified, the amount of items in the array should match the count field.
        walletSetId:
          $ref: '#/components/schemas/WalletSetId'
    BlockchainAddress:
      description: The blockchain address generated upon wallet creation.
      type: string
    WalletReferenceId:
      type: string
      description: Optional reference or description used to identify the wallet.
      example: wallet123
    WalletState:
      type: string
      enum:
        - LIVE
        - FROZEN
      example: LIVE
    WalletResponse:
      type: object
      required:
        - id
        - state
        - walletSetId
        - custodyType
        - address
        - blockchain
        - createDate
        - updateDate
      properties:
        accountType:
          type: string
          enum:
            - EOA
        address:
          $ref: '#/components/schemas/BlockchainAddress'
        blockchain:
          $ref: '#/components/schemas/Blockchain'
        createDate:
          $ref: '#/components/schemas/schemas_CreateDate'
        custodyType:
          $ref: '#/components/schemas/CustodyType'
        id:
          $ref: '#/components/schemas/schemas_Id'
        name:
          type: string
          description: Name or description associated with the wallet.
          example: My walletResponse
        refId:
          $ref: '#/components/schemas/WalletReferenceId'
        state:
          $ref: '#/components/schemas/WalletState'
        updateDate:
          $ref: '#/components/schemas/UpdateDate'
        userId:
          $ref: '#/components/schemas/ExternalUserId'
        walletSetId:
          $ref: '#/components/schemas/WalletSetId'
    CreateDeveloperWalletResponse:
      type: object
      properties:
        wallets:
          type: array
          items:
            $ref: '#/components/schemas/WalletResponse'
          description: Lists the details for the created wallets.
    ExtGetAllWalletsResponse:
      type: object
      properties:
        wallets:
          type: array
          items:
            $ref: '#/components/schemas/WalletResponse'
    ExtGetWalletByIDResponse:
      type: object
      properties:
        wallet:
          $ref: '#/components/schemas/WalletResponse'
    ExtUpdateWalletMetadataRequest:
      type: object
      properties:
        name:
          type: string
        refId:
          $ref: '#/components/schemas/WalletReferenceId'
    ExtWalletResponse:
      type: object
      properties:
        wallet:
          type: object
          properties:
            address:
              $ref: '#/components/schemas/BlockchainAddress'
            blockchain:
              $ref: '#/components/schemas/Blockchain'
            createDate:
              $ref: '#/components/schemas/schemas_CreateDate'
            custodyType:
              $ref: '#/components/schemas/CustodyType'
            id:
              $ref: '#/components/schemas/schemas_Id'
            name:
              type: string
              description: Name or description associated with the wallet.
              example: My ExtWalletResponse
            refId:
              $ref: '#/components/schemas/WalletReferenceId'
            state:
              $ref: '#/components/schemas/WalletState'
            updateDate:
              $ref: '#/components/schemas/UpdateDate'
            userId:
              $ref: '#/components/schemas/ExternalUserId'
            walletSetId:
              $ref: '#/components/schemas/WalletSetId'
    Balance:
      type: object
      required:
        - token
        - updateDate
        - amount
      properties:
        amount:
          type: string
          description: List of token balances for each token on the wallet(s).
          example: '6.62607015'
        token:
          $ref: '#/components/schemas/TokenResponse'
        updateDate:
          $ref: '#/components/schemas/UpdateDate'
    GetBalancesResponse:
      type: object
      properties:
        tokenBalances:
          description: List of token balances for each token on the wallet(s).
          type: array
          items:
            $ref: '#/components/schemas/Balance'
    Nft:
      type: object
      required:
        - nftTokenId
        - token
        - amount
        - updateDate
      properties:
        amount:
          type: string
          description: Amount of NFTs on a wallet. For non-fungible token standards, like ERC721, amount will always be “1”; for semi-fungible token standards like ERC1155, amount will correspond to the number of tokens.
          example: '1'
        metadata:
          type: string
          description: The metadata of the NFT.
          example: ipfs://QmZcH4YvBVVRJtdn4RdbaqgspFU8gH6P9vomDpBVpAL3u4/1
        nftTokenId:
          type: string
          description: The NFT token ID.
          example: '2'
        token:
          $ref: '#/components/schemas/TokenResponse'
        updateDate:
          $ref: '#/components/schemas/UpdateDate'
    GetNftsResponse:
      type: object
      properties:
        nfts:
          type: array
          items:
            $ref: '#/components/schemas/Nft'
...
    MetadataField:
      type: object
      properties:
        name:
          type: string
          description: Name or description associated with the wallet.
          example: My walletResponse
        refId:
          $ref: '#/components/schemas/WalletReferenceId'
...
    ChallengeId:
      type: string
      format: uuid
      description: Unique system generated identifier used to initiate a user challenge flow.
      example: c4d1da72-111e-4d52-bdbf-2e74a2d803d5
...
    CreateEndUserWalletRequest:
      type: object
      required:
        - blockchains
        - idempotencyKey
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
        blockchains:
          type: array
          description: Blockchain(s) the requested wallets will be created on.
          items:
            $ref: '#/components/schemas/Blockchain'
        metadata:
          type: array
          description: List of metadata fields to associate with the corresponding wallet.
          items:
            $ref: '#/components/schemas/MetadataField'
    CreateEndUserWalletResponse:
      type: object
      properties:
        challengeId:
          $ref: '#/components/schemas/ChallengeId'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PREFIX:ID:SECRET
      description: Circle's API Keys are formatted in the following structure "PREFIX:ID:SECRET". All three parts are requred to make a successful request.
  headers:
    XRequestId:
      description: Universally unique identifier (UUID v4) for the request. Helpful for identifying a request when communicating with Circle support.
      schema:
        type: string
        format: uuid
        example: 2adba88e-9d63-44bc-b975-9b6ae3440dde
  parameters:
    From:
      name: from
      description: From is the start time of the query, inclusive
      in: query
      schema:
        type: string
        format: date-time
        example: '2023-01-01T12:04:05Z'
    To:
      name: to
      description: To is the end time of the query, inclusive, default to current time
      in: query
      schema:
        type: string
        format: date-time
        example: '2023-01-01T12:04:05Z'
    PageBefore:
      name: pageBefore
      description: PageBefore is used to return items before the given item exclusively. SHOULD NOT be used in conjunction with pageAfter
      in: query
      schema:
        type: string
        format: uuid
    PageAfter:
      name: pageAfter
      description: PageAfter is used to return items after the given item exclusively. SHOULD NOT be used in conjunction with pageBefore
      in: query
      schema:
        type: string
        format: uuid
    PageSize:
      name: pageSize
      description: PageSize is the number of items to return
      in: query
      schema:
        type: integer
        default: 10
        maximum: 50
        minimum: 1
    Blockchain:
      name: blockchain
      description: Blockchain filters on the blockchain of the transaction
      in: query
      schema:
        enum:
          - ETH
          - ETH-GOERLI
          - AVAX
          - AVAX-FUJI
          - MATIC
          - MATIC-MUMBAI
        type: string
    WalletSetIdPath:
      description: WalletSet ID
      in: path
      name: id
      required: true
      schema:
        format: uuid
        type: string
    XUserToken:
      name: X-User-Token
      description: User token
      in: header
      schema:
        type: string
    Address:
      name: address
      description: Address is the address of the wallet
      in: query
      schema:
        type: string
    WalletSetIdQuery:
      name: walletSetId
      description: WalletSetID is the wallet set the wallet belongs to
      in: query
      schema:
        format: uuid
        type: string
    IncludeAll:
      name: includeAll
      in: query
      required: false
      schema:
        type: boolean
        example: true
    Name:
      name: name
      description: token name
      in: query
      schema:
        type: string
    TokenAddress:
      name: tokenAddress
      description: tokenAddress is a list of token addresses separated by comma
      in: query
      schema:
        type: string
    Standard:
      name: standard
      description: token standard
      in: query
      schema:
        type: string
...
    XUserTokenRequired:
      name: X-User-Token
      description: User token
      required: true
      in: header
      schema:
        type: string
...
  responses:
    Default:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Error response
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'